Routes.js gelöscht, in index verschoben
